move logout dropdown to bottom, move getUserAcronoym to util
